feat: add E2ETestBase for shared plumbing in end-to-end tests
